temp codes

+++ auth-server/src/main/java/org/akj/springboot/authorization/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package org.akj.springboot.authorization.repository;
 
-import org.akj.springboot.authorization.domain.User;
+import org.akj.springboot.authorization.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	/**
 	 * Search user by username
